色々 (演算子追加など)

+++ include/codegen.hpp
@@ -52,7 +52,7 @@ private:
 	llvm::Value* generateCallExpression(CallExprAST* call_expr);
 	llvm::Value* generateJumpStatement(JumpStmtAST* jump_stmt);
 	llvm::Value* generateVariable(VariableAST* var);
-	llvm::Value* generateNumber(int value);
+	llvm::Value* generateNumber(int64_t value);
 	bool linkModule(llvm::Module* dest, std::string file_name);
 };
